requirements.txt
streamlit
openai
pytesseract
pillow
pdf2image
